Update shadcn fix from monorepo-template
