+++ packages/shadcn/src/components/badge.tsx
@@ -1,28 +1,8 @@
 import { Slot } from '@radix-ui/react-slot';
 import { cn } from '@workspace/shadcn/lib/utils';
-import { cva, type VariantProps } from 'class-variance-authority';
+import { badgeVariants } from '@workspace/shadcn/variants/badge';
+import { type VariantProps } from 'class-variance-authority';
 import * as React from 'react';
-
-const badgeVariants = cva(
-  'inline-flex items-center justify-center rounded-md border px-2 py-0.5 text-xs font-medium w-fit whitespace-nowrap shrink-0 [&>svg]:size-3 gap-1 [&>svg]:pointer-events-none focus-visible:border-ring focus-visible:ring-ring/50 focus-visible:ring-[3px] aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 aria-invalid:border-destructive transition-[color,box-shadow] overflow-hidden',
-  {
-    variants: {
-      variant: {
-        default:
-          'border-transparent bg-primary text-primary-foreground [a&]:hover:bg-primary/90',
-        secondary:
-          'border-transparent bg-secondary text-secondary-foreground [a&]:hover:bg-secondary/90',
-        destructive:
-          'border-transparent bg-destructive text-white [a&]:hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:focus-visible:ring-destructive/40 dark:bg-destructive/60',
-        outline:
-          'text-foreground [a&]:hover:bg-accent [a&]:hover:text-accent-foreground',
-      },
-    },
-    defaultVariants: {
-      variant: 'default',
-    },
-  },
-);
 
 function Badge({
   className,
@@ -42,4 +22,4 @@ function Badge({
   );
 }
 
-export { Badge, badgeVariants };
+export { Badge };
