+++ packages/shadcn/src/components/button-group.tsx
@@ -1,24 +1,8 @@
 import { Slot } from '@radix-ui/react-slot';
 import { Separator } from '@workspace/shadcn/components/separator';
 import { cn } from '@workspace/shadcn/lib/utils';
-import { cva, type VariantProps } from 'class-variance-authority';
-
-const buttonGroupVariants = cva(
-  "flex w-fit items-stretch [&>*]:focus-visible:z-10 [&>*]:focus-visible:relative [&>[data-slot=select-trigger]:not([class*='w-'])]:w-fit [&>input]:flex-1 has-[select[aria-hidden=true]:last-child]:[&>[data-slot=select-trigger]:last-of-type]:rounded-r-md has-[>[data-slot=button-group]]:gap-2",
-  {
-    variants: {
-      orientation: {
-        horizontal:
-          '[&>*:not(:first-child)]:rounded-l-none [&>*:not(:first-child)]:border-l-0 [&>*:not(:last-child)]:rounded-r-none',
-        vertical:
-          'flex-col [&>*:not(:first-child)]:rounded-t-none [&>*:not(:first-child)]:border-t-0 [&>*:not(:last-child)]:rounded-b-none',
-      },
-    },
-    defaultVariants: {
-      orientation: 'horizontal',
-    },
-  },
-);
+import { buttonGroupVariants } from '@workspace/shadcn/variants/button-group';
+import { type VariantProps } from 'class-variance-authority';
 
 function ButtonGroup({
   className,
@@ -74,9 +58,4 @@ function ButtonGroupSeparator({
   );
 }
 
-export {
-  ButtonGroup,
-  ButtonGroupSeparator,
-  ButtonGroupText,
-  buttonGroupVariants,
-};
+export { ButtonGroup, ButtonGroupSeparator, ButtonGroupText };
