+++ packages/shadcn/src/components/form.tsx
@@ -3,6 +3,11 @@
 import * as LabelPrimitive from '@radix-ui/react-label';
 import { Slot } from '@radix-ui/react-slot';
 import { Label } from '@workspace/shadcn/components/label';
+import {
+  FormFieldContext,
+  FormItemContext,
+  useFormField,
+} from '@workspace/shadcn/hooks/use-form-field';
 import { cn } from '@workspace/shadcn/lib/utils';
 import * as React from 'react';
 import {
@@ -11,22 +16,9 @@ import {
   type FieldPath,
   type FieldValues,
   FormProvider,
-  useFormContext,
-  useFormState,
 } from 'react-hook-form';
 
 const Form = FormProvider;
-
-type FormFieldContextValue<
-  TFieldValues extends FieldValues = FieldValues,
-  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
-  name: TName;
-};
-
-const FormFieldContext = React.createContext<FormFieldContextValue>(
-  {} as FormFieldContextValue,
-);
 
 function FormField<
   TFieldValues extends FieldValues = FieldValues,
@@ -38,38 +30,6 @@ function FormField<
     </FormFieldContext.Provider>
   );
 }
-
-const useFormField = () => {
-  const fieldContext = React.useContext(FormFieldContext);
-  const itemContext = React.useContext(FormItemContext);
-  const { getFieldState } = useFormContext();
-  const formState = useFormState({ name: fieldContext.name });
-  const fieldState = getFieldState(fieldContext.name, formState);
-
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  if (!fieldContext) {
-    throw new Error('useFormField should be used within <FormField>');
-  }
-
-  const { id } = itemContext;
-
-  return {
-    id,
-    name: fieldContext.name,
-    formItemId: `${id}-form-item`,
-    formDescriptionId: `${id}-form-item-description`,
-    formMessageId: `${id}-form-item-message`,
-    ...fieldState,
-  };
-};
-
-type FormItemContextValue = {
-  id: string;
-};
-
-const FormItemContext = React.createContext<FormItemContextValue>(
-  {} as FormItemContextValue,
-);
 
 function FormItem({ className, ...props }: React.ComponentProps<'div'>) {
   const id = React.useId();
@@ -160,5 +120,4 @@ export {
   FormItem,
   FormLabel,
   FormMessage,
-  useFormField,
 };
