+++ packages/shadcn/src/components/navigation-menu.tsx
@@ -1,6 +1,6 @@
 import * as NavigationMenuPrimitive from '@radix-ui/react-navigation-menu';
 import { cn } from '@workspace/shadcn/lib/utils';
-import { cva } from 'class-variance-authority';
+import { navigationMenuTriggerStyle } from '@workspace/shadcn/variants/navigation-menu';
 import { ChevronDownIcon } from 'lucide-react';
 import * as React from 'react';
 
@@ -56,10 +56,6 @@ function NavigationMenuItem({
     />
   );
 }
-
-const navigationMenuTriggerStyle = cva(
-  'group inline-flex h-9 w-max items-center justify-center rounded-md bg-background px-4 py-2 text-sm font-medium hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground disabled:pointer-events-none disabled:opacity-50 data-[state=open]:hover:bg-accent data-[state=open]:text-accent-foreground data-[state=open]:focus:bg-accent data-[state=open]:bg-accent/50 focus-visible:ring-ring/50 outline-none transition-[color,box-shadow] focus-visible:ring-[3px] focus-visible:outline-1',
-);
 
 function NavigationMenuTrigger({
   className,
@@ -162,6 +158,5 @@ export {
   NavigationMenuLink,
   NavigationMenuList,
   NavigationMenuTrigger,
-  navigationMenuTriggerStyle,
   NavigationMenuViewport,
 };
